Clean up Python code, add some documentation.
Notable changes:
- Removed a bunch of code that has not been used for a long time, including unused jinja filters (perhaps most significantly: removed `conferences` from the default template context, which I'm fairly sure is unused)
- Improved detection and error reporting for rendering conference pages
- Split and moved some things into different files to simplify core.py
- Moved all scripts to _scripts
- Removed a number of obsolete files
- Added docstrings to virtually all Python code
- Renamed quite a few things to be more clear on their purpose/scope
- Extended documentation somewhat to point more in the right direction
- Cleaned up for pep8
