+++ README.md
@@ -7,24 +7,25 @@ To contribute to the Write the Docs website, it's helpful to familiarize yoursel
 
 ### Code Architecture
 
-All of the generated website lives inside the `docs` directory, but many files outside the `conf/` directory are just static RST.
-We are trying to move towards a more data driven approach,
-which allows for easier maintenance and reuse of content between events.
+All of the generated website lives inside the `docs` directory, but many files outside the `conf/` directory are just static RST, as in any other Sphinx project.
 
-#### Conference pages
+All RST files are rendered with [Jinja](http://jinja.pocoo.org/) which allows the use of Jinja tags in all of them. A few custom Jinja filters are available for things like generating photo paths for speakers.
 
-We have a few important directories and files:
+### Conference pages
+
+For conference pages, a configuration per conference is automatically loaded into the Jinja context. This allows the use of things like `{{ year }}` in RST files, and have it rendered correctly. 
+
+The important locations of files are:
 
 * `_data/config-portland-2019.yaml` - This file contains the data for rendering a specific conference.
 * `_templates/2019/` - Contains the HTML templates for all conferences that year, including common pages and separate navigation menus.
 * `conf/portland/2019/` - Contains the RST files that we use for rendering the conference. Copy these over from the *previous* conference chronologically, not the previous conference in the same location.
 * `include/conf/` Contains the text snippets which are *mostly* the same between all conferences, mostly describing what our events are (eg. lightning talks and unconference). They are included via the RST files for each event. Now we've merged #738 we can add content that is conditional to each conference using Jinja.
-* `_ext/core.py` Contains the Sphinx extensions that manage injecting custom variables into our RST and Jinja templates. Specifically the `rstjinja` and `load_page_yaml_data` functions do a lot of the work.
+* `_ext/core.py` Contains the Sphinx extensions that manage injecting custom variables into our RST and Jinja templates. Specifically the `render_rst_with_jinja` and `load_conference_page_context` functions do a lot of the work.
 
-All files that live under the `conf` directory are rendered so as to have access to
- the data inside the `_data/config-<location>-<year>.yaml` file in the Jinja context for the RST files.
-This allows us to say `{{ year }}` in the RST files,
-and have it be rendered properly at `2019`.
+The YAML config file's name must match the year and shortcode of the conference, i.e. for `conf/portland/2019`, the YAML loaded is `_data/config-portland-2019`.
+
+Each conference also has YAML files for the schedule and speakers, in the format of `_data/2019.portland.speakers.yaml` and `_data/portland-2019-day-1.yaml`. These are rendered using Sphinx datatemplates, and also used in processes like generating video pages (a manual task, not done every time while building).
 
 ### Videos
 
@@ -62,9 +63,8 @@ If you run into trouble with broken links to video files, have a look at `_ext/f
 
 ### Prerequisites for generating the docs locally
 
-You'll probably need `root` privileges to install the prerequisites.
-
-1. Install `python 3.6.x` using your package manager.
+1. Install `python 3.6.x` using your package manager, if not installed already.
+   You'll probably need `root` privileges to do this.
 
 2. Generate a virtual environment for the WTD repo in the `venv` directory:
