+++ docs/_ext/videos.py
@@ -9,13 +9,51 @@ Create new videos by adding the following to ``conf.py``::
 """
 from __future__ import print_function
 
+import glob
+import re
+
 from builtins import str
 import io
 import os
 import os.path
-import shutil
 
-from .core import generate_video_slug, load_conference_data
+from _ext.utils import generate_video_slug, load_yaml, normalize_session
+
+
+def load_conference_data():
+    """
+    Generate a dict with all data of all conferences including
+    session details.
+    """
+    speakers_file_pattern = re.compile(r'(\d{4}).(\w+).speakers')
+    sessions_file_pattern = re.compile(r'(\w+)-(\d{4})-day-(\d+)')
+    result = {}
+    for f in glob.glob('_data/*.yaml'):
+        base = os.path.basename(f)
+        base, _ = os.path.splitext(base)
+        # Only consider conference data files that are following the common
+        # naming convention and log everything out as warning.
+        mo = speakers_file_pattern.match(base)
+        if mo:
+            year = int(mo.group(1))
+            region = mo.group(2)
+            if year not in result:
+                result[year] = {}
+            if region not in result[year]:
+                result[year][region] = {}
+            result[year][region]['speakers'] = load_yaml(f)
+            for session in result[year][region]['speakers']:
+                normalize_session(session)
+                session['year'] = year
+                session['series'] = u'Write the Docs {}'.format(region.upper())
+                session['series_slug'] = region
+                session['event'] = u'Write the Docs {} {}'.format(region.upper(), year)
+                session['path'] = 'conf/{series_slug}/{year}/videos/{slug}'.format(**session)
+            continue
+        mo = sessions_file_pattern.match(base)
+        if mo:
+            continue
+    return result
 
 
 def generate_video_content(session, year, region, session_idx):
