+++ docs/_scripts/make-email.py
@@ -24,13 +24,11 @@ url = sys.argv[1]
 
 d = pq(url=url)
 content = d('.col-content').html()
-# import ipdb; ipdb.set_trace();
 if not content:
     content = d('.page-content').html()
     # Remove page title
     d = pq(content)
     d.find('span').remove()
-    #d.find('h1').remove()
     # Convert images
     for img_obj in d('img'):
         img = pq(img_obj)
@@ -43,7 +41,7 @@ if not content:
 # Remove header links
 try:
     content = re.sub(r'<a class="headerlink" .+</a>', '', content)
-except:
+except Exception:
     pass
